[CHORE] #106 : 출석 관리 다이얼로그 Component 분리
